deinline n_poly IO
