+++ src/n_poly/n_polyu.c
@@ -53,31 +53,6 @@ void n_polyu_realloc(n_polyu_t A, slong len)
     A->alloc = new_alloc;
 }
 
-void n_polyu3_print_pretty(
-    const n_polyu_t A,
-    const char * var0,
-    const char * var1,
-    const char * var2)
-{
-    slong i;
-    int first = 1;
-
-    for (i = 0; i < A->length; i++)
-    {
-        if (!first)
-            printf(" + ");
-        first = 0;
-        flint_printf("%wu*%s^%wu*%s^%wu*%s^%wu",
-            A->coeffs[i],
-            var0, extract_exp(A->exps[i], 2, 3),
-            var1, extract_exp(A->exps[i], 1, 3),
-            var2, extract_exp(A->exps[i], 0, 3));
-    }
-
-    if (first)
-        flint_printf("0");
-}
-
 void n_polyu3_degrees(
     slong * deg0,
     slong * deg1,
