+++ src/n_poly/n_polyun.c
@@ -61,81 +61,6 @@ void n_polyun_realloc(n_polyun_t A, slong len)
     A->alloc = new_alloc;
 }
 
-void n_polyu1n_print_pretty(
-    const n_polyun_t A,
-    const char * var0,
-    const char * varlast)
-{
-    slong i;
-    int first = 1;
-
-    for (i = 0; i < A->length; i++)
-    {
-        if (!first)
-            flint_printf(" + ");
-        first = 0;
-        flint_printf("(");
-        n_poly_print_pretty(A->coeffs + i, varlast);
-        flint_printf(")*%s^%wu", var0, A->exps[i]);
-    }
-
-    if (first)
-        flint_printf("0");
-}
-
-void n_polyu2n_print_pretty(
-    const n_polyun_t A,
-    const char * var0,
-    const char * var1,
-    const char * varlast)
-{
-    slong i;
-    int first = 1;
-
-    for (i = 0; i < A->length; i++)
-    {
-        if (!first)
-            flint_printf(" + ");
-        first = 0;
-        flint_printf("(");
-        n_poly_print_pretty(A->coeffs + i, varlast);
-        flint_printf(")*%s^%wu*%s^%wu",
-            var0, extract_exp(A->exps[i], 1, 2),
-            var1, extract_exp(A->exps[i], 0, 2));
-    }
-
-    if (first)
-        flint_printf("0");
-}
-
-void n_polyu3n_print_pretty(
-    const n_polyun_t A,
-    const char * var0,
-    const char * var1,
-    const char * var2,
-    const char * varlast)
-{
-    slong i;
-    int first = 1;
-
-    for (i = 0; i < A->length; i++)
-    {
-        if (!first)
-            flint_printf(" + ");
-        first = 0;
-        flint_printf("(");
-        n_poly_print_pretty(A->coeffs + i, varlast);
-        flint_printf(")*%s^%wu*%s^%wu*%s^%wu",
-            var0, extract_exp(A->exps[i], 2, 3),
-            var1, extract_exp(A->exps[i], 1, 3),
-            var2, extract_exp(A->exps[i], 0, 3));
-    }
-
-    if (first)
-        flint_printf("0");
-}
-
-
 void n_polyun_set(n_polyun_t A, const n_polyun_t B)
 {
     slong i;
